+++ supabase/schemas/tokens.sql
@@ -65,7 +65,7 @@ with
 		)
 	);
 
-create policy "only token owners and venue admins can update a token" on public.tokens
+create policy "only token owners, venue admins, and minters can update a token" on public.tokens
 for update
 	to authenticated using (
 		(
@@ -80,6 +80,21 @@ for update
 						select
 							auth.uid () as uid
 					)=scholar
+				)
+			)
+			or (
+				(
+					select
+						auth.uid () as uid
+				)=any (
+					(
+						select
+							currencies.minters
+						from
+							public.currencies
+						where
+							(currencies.id=tokens.currency)
+					)::uuid[]
 				)
 			)
 		)